requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.1
joblib==1.4.2
Pillow==10.4.0